docs: Documentación del BuscadorDeTarea, OrganizadorDeTarea y ListadoDeTarea que antes era coleccionDeTarea.
